generador/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==0.8.11
gunicorn==21.2.0